Update stop watch to use a custom hook, refactor directory structure
