+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import StopWatch from './StopWatch';
+import StopWatch from './components/StopWatch';
 import '../styles/StopWatchButton.css';
 import '../styles/StopWatch.css';
