+++ src/utils/FormatTime.tsx
@@ -1,4 +1,3 @@
-
 // Function to add a leading zero given a number
 // ex: num = 1
 //     return = 01
@@ -7,7 +6,6 @@
 const addLeadingZero = (num: number): string => {
     return `${num < 10 ? '0' : ''}${num}`
 }
-
 
 export const formatTime = (timeInSeconds: number): string => {
     const hours: number = Math.floor(timeInSeconds / 3600);
