+++ src/utils/TimeUtils.tsx
@@ -1,0 +1,3 @@
+export const getCurrentTime = (): number => {
+    return Date.now();
+}
